requirements.txt
django==3.2.7
django-extensions
ipython
django-debug-toolbar
django-import-export
django-admin-rangefilter
django-silk
django-storages
django-crispy-forms
boto3
celery
requests
beautifulsoup4
lxml
django-filter

# flake8
flake8
flake8-builtins
flake8-import-order
flake8-print
flake8-breakpoint